Refactor(mentor): 서비스통합에 따른 경로 변경
