+++ pages/articles/[id]/index.js
@@ -2,8 +2,7 @@ import Image from 'next/image';
 import { useState, useRef } from 'react';
 import { useRouter } from 'next/router';
 import styles from '@/styles/Article.module.css';
-import { getArticle } from '@/lib/api/ArticleService';
-import { getArticleCommentList } from '@/lib/api/ArticleCommentService';
+import { getArticle, getArticleCommentList } from '@/lib/api/ArticleService';
 import ArticleCommentAdd from '@/components/ArticleDetail/ArticleCommentAdd';
 import ArticleCommentList from '@/components/ArticleDetail/ArticleCommentList';
 import formatDate from '@/lib/formatDate';
